Support db urls without port and username/password.
The database url had to include port, username and password, like this:
`jdbc:postgresql://localhost:5432/mydb?username=foo&password=bar`

This change allows to leave out port or username/password, like this:
`jdbc:postgresql://localhost/mydb`
(for simplicity only both username and password can be left out, not
username or password)

This is implemented for `jdbc:postgresql` protocol, not for heroku
`postgresql`/`postgres` protocol.

This commit also includes a change of the default port from 5234 to 5432.

Closes #32