change the default utf8 encoding
according to this post https://mathiasbynens.be/notes/mysql-utf8mb4
the default utf8 in mysql do not exactly match utf8 standard (3 vs 4 bytes)

Since it maps to the same charset in Java
It sounds like the reasonable thing to do is to change the used protocol
from 83 to 224 (utf8mb4_unicode_ci)
I added a flag to allow overriding it `-DjasyncMysqlUTF8Collation=83`
To put in back in the old legacy behaviour
for example if using mysql server < 5.5.3.

Added also tests for this.