Support SSL for MySQL (#222)
This PR adds support for the ssl configuration in the MySQL driver. There is also a refactoring to keep the SSL related code shared between the drivers (because the drivers are doing the same START_TLS mechanism).

The SSL guide in the protocol can be found here: https://dev.mysql.com/doc/internals/en/ssl.html

We add a new message for SSLRequest that we send if the config requests SSL and the server supports it. Once we get the handshake from the server we send this request then upgrade to a SSL connection, then send the handshake response as usual.

The SSLRequest message is the same as the handshake response message, just truncated after the padding. I refactored the handshake response message to share that code. I'm happy to change this bit up if you'd rather not share that bit (this was inspired from the way other drivers model this), but the spec doesn't talk about this "header" concept so I can see the case for sharing this serialization in another way.