fix: futures in waiting queue should also timeout if can't create items (#225)
In cases that the pool can't create new objects it will cancel futures
already waiting. But new futures that joins the queue will stay there
forever without respecting any timeout.
This fix check also the waiting items for timeout. The queryTimeout is
used and not create timeout because when pool is loaded futures might be
timeout even when there are active items. That is still more predictable
than waiting in the queue more than the query timeout.
In some edge cases queries might still be timeout after queryTimeout * 2
- when pool is loaded and it almost hit the queryTimeout in wait and
started just before that, and then hitting a timeout on the query
itself. This is because the queryTimeout is not accumulated between the
wait time and the execution time, and it doesn't looks like worth the
effort to change that at the moment - as this is really an edge case and
it still causes a timeout eventually.

Fixes #223