an initial impl for r2dbc filter/flatmap
on operation
create a new instance from result `JasyncSegmentResult`
JasyncSegmentResult has a flux of segments which are
rows + number of rows segment.
Then filter and flatmap are operating on that flux.

There is no error handling here as the original result
don't have such concept.
JasyncStatement.execute() either yields result or exception (exception is not a result).

see #351